% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folg_plays.R
\name{folg_plays}
\alias{folg_plays}
\title{Shakespeare Play Codes}
\usage{
folg_plays()
}
\value{
a named character vector, whose entry names are the titles of
Shakespeare's plays, and values are the corresponding play codes.
}
\description{
Returns the codes of Shakespeare's plays available in the Folger
Shakespeare Library.
}
\examples{
folg_plays()
}
\author{
Valerio Gherardi
}
