% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folg_synopsis.R
\name{folg_synopsis}
\alias{folg_synopsis}
\title{Spoken Text in a Play}
\usage{
folg_synopsis(play)
}
\arguments{
\item{play}{a length one character. Code of the required play, i.e. one of
the strings returned by \code{\link{folg_plays}()}.}
}
\value{
a list of two elements:
\itemize{
\item \code{summary}: a length one character vector, containing a summary of the
play.
\item \code{scenes}: a character vector, whose entries are summaries of the
singular scenes of the play.
}
}
\description{
Returns a synopsis of the play and its scenes
}
\examples{
folg_synopsis("AWW")
}
\author{
Valerio Gherardi
}
