% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{folg_text}
\alias{folg_text}
\title{Spoken Text in a Play}
\usage{
folg_text(play)
}
\arguments{
\item{play}{a length one character. Code of the required play, i.e. one of
the strings returned by \code{\link{folg_plays}()}.}
}
\value{
a character vector.
}
\description{
Returns only the spoken text in the selected play.
}
\examples{
folg_text("AWW")
}
\author{
Valerio Gherardi
}
