% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folgeR-package.R
\docType{package}
\name{folgeR-package}
\alias{folgeR}
\alias{folgeR-package}
\title{folgeR: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vgherard/folgeR}
  \item Report bugs at \url{https://github.com/vgherard/folgeR/issues}
}

}
\keyword{internal}
