% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folgeR-package.R
\docType{package}
\name{folgeR-package}
\alias{folgeR}
\alias{folgeR-package}
\title{folgeR: Client for the Folger Shakespeare Library API}
\description{
Client for the Folger Shakespeare Library API.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vgherard/folgeR}
  \item Report bugs at \url{https://github.com/vgherard/folgeR/issues}
}

}
\author{
\strong{Maintainer}: Valerio Gherardi \email{vgherard@sissa.it} (\href{https://orcid.org/0000-0002-8215-3013}{ORCID})

}
\keyword{internal}
